requirements.txt
winotify
pyyaml